+++ src/contentDisplayer.ts
@@ -1,0 +1,4 @@
+import * as vscode from 'vscode';
+export interface ContentDisplayer {
+    displayAsReadOnly(uri:vscode.Uri, readFromCache?:boolean):Promise<vscode.TextEditor>;
+}