+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.16] - 2018-??-??
+- (issue 53) Added support for campaign overrides.
+- (issue 54) Lytics content provider read directly from Lytics every time (instead of using the content provider cache).
+- Refactored content explorers to use inheritance to reduce code duplication.
+
 ## [0.0.15] - 2018-08-20
 - (issue 43) Fixed error that was displayed when loading a Lytics account with no campaigns.
 - (issue 44) Queries are sorted alphabetically.